OpenGraph as Component
