+++ components/index.js
@@ -23,4 +23,5 @@ export { Disqus } from "./Disqus";
 
 export { Layout } from "./Layout";
 export { MDXComponents } from "./MDXComponents";
+export { OpenGraph } from "./OpenGraph";
 export * as elements from "./elements";
